Refractor : Add tests to the same file
